Make README less misleading
The previous version of the README could be
interpreted in a way that Shallows perform
some kind of a deep search inside the contents
of the structure. This commit changes the order
of the lines to show that in order for the
`retrieve` to be succesfull, the value should
be `set` previously.